refactor: modularize frontend architecture, extract auth and dashboard views
